Fixed multiple bugs that were the result of implicit conversion

+++ k_means.c
@@ -1,17 +1,18 @@
 #include <math.h>
-#include <limits.h>
+#include <float.h>
 #include <stdio.h>
 /*Returns the square of the euclidean distance between two 2D points. Since the only thing that is important in K-Means clustering is relative disnce, i.e. is this point further from this center or this center, and not the actual distance we don't take the root of this number because calling sqrt() would be expensive.*/
 //TODO make this n-dimensional
-int distance(double p1[], double p2[]){
-	return pow(p1[0]-p2[0], 2)+pow(p1[1]-p2[1], 2);
+double distance(double p1[], double p2[]){
+	return pow(p1[0]-p2[0],2) + pow(p1[1]-p2[1], 2);
 }
 
 double k_means(int k, double** centers, double** dataset,  __int8_t** clusters, int maxindex){
 	double point[2], center[2], minDistance, minDistanceCenter, totalDistance, curDistance;
 	int pointIndex, centerIndex, minDistanceIndex, i;
 	for(pointIndex=0; pointIndex<maxindex; pointIndex++){
-		minDistance=INT_MAX;
+		
+		minDistance=DBL_MAX;
 		point[0]=dataset[0][pointIndex];
 		point[1]= dataset[0][pointIndex];
 		for(centerIndex=0; centerIndex<k; centerIndex++){
